fix sonarlint issues

+++ src/parallel.cpp
@@ -1,37 +1,35 @@
 #include "parallel.hpp"
-#include <functional>
 #include <thread>
 #include <vector>
 
-void parallel::for_each(unsigned nb_elements,
-                        std::function<void(int start, int end)> functor,
+void parallel::for_each(size_t nb_elements,
+                        std::function<void(size_t start, size_t end)> functor,
                         bool use_threads) {
-  // -------
-  unsigned nb_threads_hint = std::thread::hardware_concurrency();
-  unsigned nb_threads = nb_threads_hint == 0 ? 8 : (nb_threads_hint);
+  size_t nb_threads_hint = std::thread::hardware_concurrency();
+  size_t nb_threads = nb_threads_hint == 0 ? 8 : nb_threads_hint;
 
-  unsigned batch_size = nb_elements / nb_threads;
-  unsigned batch_remainder = nb_elements % nb_threads;
+  size_t batch_size = nb_elements / nb_threads;
+  size_t batch_remainder = nb_elements % nb_threads;
 
   std::vector<std::jthread> my_threads(nb_threads);
 
   if (use_threads) {
     // Multithread execution
     for (unsigned i = 0; i < nb_threads; ++i) {
-      int start = i * batch_size;
+      size_t start = i * batch_size;
       my_threads[i] = std::jthread(functor, start, start + batch_size);
     }
   } else {
     // Single thread execution (for easy debugging)
     for (unsigned i = 0; i < nb_threads; ++i) {
-      int start = i * batch_size;
+      size_t start = i * batch_size;
       functor(start, start + batch_size);
     }
   }
 
   // Deform the elements left
-  int start = nb_threads * batch_size;
+  size_t start = nb_threads * batch_size;
   functor(start, start + batch_remainder);
 
-  // jthread gets automatically awaited when destructed
+  // jthreads are automatically avaited at the end of the scope
 }
