+++ src/parallel.hpp
@@ -1,4 +1,4 @@
-// from https://stackoverflow.com/a/49188371
+// adapted from https://stackoverflow.com/a/49188371
 
 #include <functional>
 
@@ -16,8 +16,8 @@ namespace parallel {
 /// @param use_threads : enable / disable threads.
 ///
 ///
-void for_each(unsigned nb_elements,
-              std::function<void(int start, int end)> functor,
+void for_each(size_t nb_elements,
+              std::function<void(size_t start, size_t end)> functor,
               bool use_threads = true);
 
 } // namespace parallel
